tests/test_grid_calculator.py
"""
Unit tests for Grid Bot calculations and logic.

Tests cover:
- Grid level calculation with arithmetic spacing
- Price/quantity rounding to valid tick/lot sizes
- Order placement logic
"""
import pytest
from decimal import Decimal
from unittest.mock import AsyncMock, MagicMock, patch

# Import will fail without dependencies, but we test logic independently
import sys
sys.path.insert(0, '..')


class TestGridCalculation:
    """Test grid level calculation."""
    
    def test_arithmetic_grid_spacing(self):
        """
        Test that grid levels are evenly spaced (arithmetic progression).
        
        Formula: grid_step = (upper - lower) / (count - 1)
        """
        lower = Decimal("0.9200")
        upper = Decimal("1.0200")
        count = 11  # 11 levels = 10 spaces
        
        expected_step = (upper - lower) / (count - 1)  # 0.01
        
        # Generate levels
        levels = []
        for i in range(count):
            price = lower + (Decimal(i) * expected_step)
            levels.append(price)
        
        # Verify spacing is consistent
        for i in range(1, len(levels)):
            diff = levels[i] - levels[i-1]
            assert diff == expected_step, f"Spacing at index {i} is {diff}, expected {expected_step}"
        
        # Verify boundaries
        assert levels[0] == lower
        assert levels[-1] == upper
    
    def test_grid_range_from_percent(self):
        """Test grid range calculation from percentage."""
        current_price = Decimal("0.9683")
        range_percent = Decimal("5.0")  # ±5%
        
        lower = current_price * (1 - range_percent / 100)
        upper = current_price * (1 + range_percent / 100)
        
        expected_lower = Decimal("0.919885")  # 0.9683 * 0.95
        expected_upper = Decimal("1.016715")  # 0.9683 * 1.05
        
        assert abs(lower - expected_lower) < Decimal("0.000001")
        assert abs(upper - expected_upper) < Decimal("0.000001")
    
    def test_grid_order_sides(self):
        """Test that BUY orders are below current price, SELL above."""
        current_price = Decimal("0.9683")
        
        test_prices = [
            (Decimal("0.9500"), "BUY"),
            (Decimal("0.9600"), "BUY"),
            (Decimal("0.9700"), "SELL"),
            (Decimal("0.9800"), "SELL"),
        ]
        
        for price, expected_side in test_prices:
            if price < current_price:
                actual_side = "BUY"
            else:
                actual_side = "SELL"
            
            assert actual_side == expected_side, f"Price {price}: expected {expected_side}, got {actual_side}"


class TestPriceRounding:
    """Test price and quantity rounding to valid exchange values."""
    
    def test_price_round_to_tick(self):
        """Test price rounding to tick size."""
        tick_size = Decimal("0.0001")
        
        test_cases = [
            (Decimal("0.96834567"), Decimal("0.9683")),  # Round down
            (Decimal("0.96839999"), Decimal("0.9683")),  # Round down
            (Decimal("0.9683"), Decimal("0.9683")),     # Exact
        ]
        
        from decimal import ROUND_DOWN
        
        for input_price, expected in test_cases:
            rounded = (input_price / tick_size).quantize(Decimal("1"), ROUND_DOWN) * tick_size
            assert rounded == expected, f"Input {input_price}: expected {expected}, got {rounded}"
    
    def test_quantity_round_to_lot(self):
        """Test quantity rounding to lot size."""
        lot_size = Decimal("0.01")
        
        test_cases = [
            (Decimal("10.567"), Decimal("10.56")),  # Round down
            (Decimal("10.999"), Decimal("10.99")),  # Round down
            (Decimal("10.01"), Decimal("10.01")),   # Exact
        ]
        
        from decimal import ROUND_DOWN
        
        for input_qty, expected in test_cases:
            rounded = (input_qty / lot_size).quantize(Decimal("1"), ROUND_DOWN) * lot_size
            assert rounded == expected, f"Input {input_qty}: expected {expected}, got {rounded}"


class TestQuantityCalculation:
    """Test order quantity calculation."""
    
    def test_quantity_from_usdt_value(self):
        """
        Test converting USDT value to base asset quantity.
        
        Formula: quantity = (usdt_per_grid * leverage) / price
        """
        usdt_per_grid = Decimal("50.0")
        leverage = Decimal("2")
        price = Decimal("0.9683")
        
        expected_quantity = (usdt_per_grid * leverage) / price  # ~103.28
        
        assert expected_quantity > Decimal("103")
        assert expected_quantity < Decimal("104")
    
    def test_minimum_notional_enforcement(self):
        """Test that orders meet minimum notional value."""
        min_notional = Decimal("5")
        price = Decimal("0.9683")
        
        # Calculate minimum quantity to meet notional
        min_quantity = min_notional / price  # ~5.16
        
        assert min_quantity * price >= min_notional


class TestDrawdownCalculation:
    """Test drawdown percentage calculation."""
    
    def test_no_drawdown_on_profit(self):
        """Drawdown should be 0 when in profit."""
        initial_balance = Decimal("500")
        current_balance = Decimal("520")
        unrealized_pnl = Decimal("10")
        
        current_equity = current_balance + unrealized_pnl
        pnl = current_equity - initial_balance
        
        if pnl >= 0:
            drawdown = Decimal("0")
        else:
            drawdown = abs(pnl) / initial_balance * 100
        
        assert drawdown == Decimal("0")
    
    def test_drawdown_percentage_calculation(self):
        """Test drawdown percentage on loss."""
        initial_balance = Decimal("500")
        current_balance = Decimal("450")
        unrealized_pnl = Decimal("-20")
        
        current_equity = current_balance + unrealized_pnl  # 430
        pnl = current_equity - initial_balance  # -70
        
        expected_drawdown = abs(pnl) / initial_balance * 100  # 14%
        
        assert expected_drawdown == Decimal("14")


class TestHMACSignature:
    """Test HMAC-SHA256 signature generation."""
    
    def test_signature_generation(self):
        """Test HMAC signature matches expected output."""
        import hmac
        import hashlib
        from urllib.parse import urlencode
        
        secret = "test_secret_key"
        params = {
            "symbol": "ASTERUSDT",
            "side": "BUY",
            "type": "LIMIT",
            "timestamp": 1702252800000,
        }
        
        # Create sorted query string
        query_string = urlencode(sorted(params.items()))
        
        # Compute signature
        signature = hmac.new(
            secret.encode("utf-8"),
            query_string.encode("utf-8"),
            hashlib.sha256
        ).hexdigest()
        
        # Signature should be 64 character hex string
        assert len(signature) == 64
        assert all(c in "0123456789abcdef" for c in signature)
    
    def test_signature_consistency(self):
        """Test that same inputs produce same signature."""
        import hmac
        import hashlib
        from urllib.parse import urlencode
        
        secret = "secret"
        params = {"a": "1", "b": "2"}
        query_string = urlencode(sorted(params.items()))
        
        sig1 = hmac.new(secret.encode(), query_string.encode(), hashlib.sha256).hexdigest()
        sig2 = hmac.new(secret.encode(), query_string.encode(), hashlib.sha256).hexdigest()
        
        assert sig1 == sig2


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
