requirements.txt
# Core dependencies
aiohttp>=3.9.0
websockets>=12.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# Data Analysis (for StrategyManager)
pandas>=2.1.0
numpy>=1.26.0

# Database (optional, but good practice if switching from sqlite)
# asyncpg>=0.29.0  # For PostgreSQL on Railway later

# Optimization (optional)
uvloop>=0.19.0; sys_platform != "win32"
